+++ sublime/cache_manager.py
@@ -1,9 +1,11 @@
 import os
 import glob
+import itertools
 import threading
 import shutil
 import json
 import hashlib
+
 from collections import defaultdict
 from time import sleep
 
@@ -26,6 +28,8 @@ from typing import (
 
 import requests
 
+from fuzzywuzzy import fuzz
+
 from .config import AppConfiguration, ServerConfiguration
 from .server import Server
 from .server.api_object import APIObject
@@ -44,6 +48,7 @@ from .server.api_objects import (
     ArtistID3,
     ArtistInfo2,
     ArtistWithAlbumsID3,
+    AlbumID3,
     AlbumWithSongsID3,
 )
 
@@ -68,6 +73,81 @@ class SongCacheStatus(Enum):
     CACHED = 1
     PERMANENTLY_CACHED = 2
     DOWNLOADING = 3
+
+
+class SearchResult:
+    _artist: Set[Union[Artist, ArtistID3]] = set()
+    _album: Set[Union[Child, AlbumID3]] = set()
+    _song: Set[Child] = set()
+    _playlist: Set[Playlist] = set()
+
+    def __init__(self, query):
+        self.query = query
+
+    def add_results(self, result_type, results):
+        if results is None:
+            return
+
+        member = f'_{result_type}'
+        if getattr(self, member) is None:
+            setattr(self, member, set())
+
+        setattr(
+            self,
+            member,
+            getattr(getattr(self, member, set()), 'union')(set(results)),
+        )
+
+    def _to_result(self, it, transform):
+        all_results = sorted(
+            (
+                (
+                    fuzz.partial_ratio(
+                        self.query.lower(),
+                        transform(x).lower(),
+                    ),
+                    x,
+                ) for x in it),
+            key=lambda rx: rx[0],
+            reverse=True,
+        )
+        result = []
+        for ratio, x in all_results:
+            if ratio > 60 and len(result) < 20:
+                result.append(x)
+            else:
+                # No use going on, all the rest are less.
+                break
+        return result
+
+    @property
+    def artist(self) -> Optional[List[Union[Artist, ArtistID3]]]:
+        if self._artist is None:
+            return None
+        return self._to_result(self._artist, lambda a: a.name)
+
+    @property
+    def album(self) -> Optional[List[Union[Child, AlbumID3]]]:
+        if self._album is None:
+            return None
+
+        def album_transform(a):
+            name = a.title if type(a) == Child else a.name
+            return f'{name} - {a.artist}'
+
+        return self._to_result(self._album, album_transform)
+
+    @property
+    def song(self) -> Optional[List[Child]]:
+        if self._song is None:
+            return None
+        return self._to_result(self._song, lambda s: f'{s.title} - {s.artist}')
+
+    @property
+    def playlist(self) -> Optional[List[Playlist]]:
+        if self._playlist is None:
+            return None
+        return self._to_result(self._playlist, lambda p: p.name)
 
 
 T = TypeVar('T')
@@ -179,9 +259,9 @@ class CacheManager(metaclass=Singleton):
         current_downloads: Set[str] = set()
 
         def __init__(
-                self,
-                app_config: AppConfiguration,
-                server_config: ServerConfiguration,
+            self,
+            app_config: AppConfiguration,
+            server_config: ServerConfiguration,
         ):
             self.app_config = app_config
             self.browse_by_tags = self.app_config.server.browse_by_tags
@@ -213,7 +293,6 @@ class CacheManager(metaclass=Singleton):
                 # Playlists
                 ('playlists', Playlist, list),
                 ('playlist_details', PlaylistWithSongs, dict),
-                ('song_details', Child, dict),
                 ('genres', Genre, list),
 
                 # Non-ID3 caches
@@ -222,6 +301,7 @@ class CacheManager(metaclass=Singleton):
                 ('artists', Artist, list),
                 ('artist_details', Directory, dict),
                 ('artist_infos', ArtistInfo, dict),
+                ('song_details', Child, dict),
 
                 # ID3 caches
                 ('albums_id3', AlbumWithSongsID3, 'dict-list'),
@@ -229,6 +309,7 @@ class CacheManager(metaclass=Singleton):
                 ('artists_id3', ArtistID3, list),
                 ('artist_details_id3', ArtistWithAlbumsID3, dict),
                 ('artist_infos_id3', ArtistInfo2, dict),
+                ('song_details_id3', Child, dict),
             ]
             for name, type_name, default in cache_configs:
                 if default == list:
@@ -283,12 +364,12 @@ class CacheManager(metaclass=Singleton):
             )
 
         def return_cached_or_download(
-                self,
-                relative_path: Union[Path, str],
-                download_fn: Callable[[], bytes],
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
-                allow_download: bool = True,
+            self,
+            relative_path: Union[Path, str],
+            download_fn: Callable[[], bytes],
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
+            allow_download: bool = True,
         ) -> 'CacheManager.Result[Optional[str]]':
             abs_path = self.calculate_abs_path(relative_path)
             abs_path_str = str(abs_path)
@@ -358,9 +439,9 @@ class CacheManager(metaclass=Singleton):
                 Path(path).unlink()
 
         def get_playlists(
-                self,
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
+            self,
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
         ) -> 'CacheManager.Result[List[Playlist]]':
             if self.cache.get('playlists') and not force:
                 return CacheManager.Result.from_data(self.cache['playlists'])
@@ -385,10 +466,10 @@ class CacheManager(metaclass=Singleton):
             self.save_cache_info()
 
         def get_playlist(
-                self,
-                playlist_id: int,
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
+            self,
+            playlist_id: int,
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
         ) -> 'CacheManager.Result[PlaylistWithSongs]':
             playlist_details = self.cache.get('playlist_details', {})
             if playlist_id in playlist_details and not force:
@@ -401,7 +482,7 @@ class CacheManager(metaclass=Singleton):
 
                     # Playlists have the song details, so save those too.
                     for song in (playlist.entry or []):
-                        self.cache['song_details'][song.id] = song
+                        self.cache[self.id3ify('song_details')][song.id] = song
 
                 self.save_cache_info()
 
@@ -431,9 +512,9 @@ class CacheManager(metaclass=Singleton):
             return CacheManager.create_future(do_update_playlist)
 
         def get_artists(
-                self,
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
+            self,
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
         ) -> 'CacheManager.Result[List[Union[Artist, ArtistID3]]]':
             cache_name = self.id3ify('artists')
 
@@ -463,10 +544,10 @@ class CacheManager(metaclass=Singleton):
             )
 
         def get_artist(
-                self,
-                artist_id,
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
+            self,
+            artist_id,
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
         ) -> 'CacheManager.Result[Union[ArtistWithAlbumsID3, Child]]':
             cache_name = self.id3ify('artist_details')
 
@@ -490,10 +571,10 @@ class CacheManager(metaclass=Singleton):
             )
 
         def get_artist_info(
-                self,
-                artist_id,
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
+            self,
+            artist_id,
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
         ) -> 'CacheManager.Result[Union[ArtistInfo, ArtistInfo2]]':
             cache_name = self.id3ify('artist_infos')
 
@@ -526,13 +607,13 @@ class CacheManager(metaclass=Singleton):
                 force: bool = False,
         ) -> 'CacheManager.Result[Optional[str]]':
             def do_get_artist_artwork(artist_info):
-                lastfm_url = ''.join(artist_info.largeImageUrl)
+                lastfm_url = ''.join(artist_info.largeImageUrl or [])
 
                 # If it is the placeholder LastFM image, try and use the cover
                 # art filename given by the server.
                 if (lastfm_url == '' or lastfm_url.endswith(
                         '2a96cbd8b46e442fc41c2b86b821562f.png')):
-                    if isinstance(artist, ArtistWithAlbumsID3):
+                    if isinstance(artist, (ArtistWithAlbumsID3, ArtistID3)):
                         return CacheManager.get_cover_art_filename(
                             artist.coverArt, size=300)
                     elif (isinstance(artist, Directory)
@@ -540,6 +621,9 @@ class CacheManager(metaclass=Singleton):
                         # Retrieve the first album's cover art
                         return CacheManager.get_cover_art_filename(
                             artist.child[0].coverArt, size=300)
+
+                if lastfm_url == '':
+                    return CacheManager.Result.from_data('')
 
                 url_hash = hashlib.md5(lastfm_url.encode('utf-8')).hexdigest()
                 return self.return_cached_or_download(
@@ -565,12 +649,12 @@ class CacheManager(metaclass=Singleton):
                 return do_get_artist_artwork(artist_info.result())
 
         def get_album_list(
-                self,
-                type_: str,
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
-                # Look at documentation for get_album_list in server.py:
-                **params,
+            self,
+            type_: str,
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
+            # Look at documentation for get_album_list in server.py:
+            **params,
         ) -> 'CacheManager.Result[List[Union[Child, AlbumWithSongsID3]]]':
             cache_name = self.id3ify('albums')
 
@@ -628,10 +712,10 @@ class CacheManager(metaclass=Singleton):
             self.save_cache_info()
 
         def get_album(
-                self,
-                album_id,
-                before_download: Callable[[], None] = lambda: None,
-                force: bool = False,
+            self,
+            album_id,
+            before_download: Callable[[], None] = lambda: None,
+            force: bool = False,
         ) -> 'CacheManager.Result[Union[AlbumWithSongsID3, Child]]':
             cache_name = self.id3ify('album_details')
 
@@ -646,6 +730,11 @@ class CacheManager(metaclass=Singleton):
             def after_download(album):
                 with self.cache_lock:
                     self.cache[cache_name][album_id] = album
+
+                    # Albums have the song details as well, so save those too.
+                    for song in (album.get('song') or album.get('child')
+                                 or []):
+                        self.cache[self.id3ify('song_details')][song.id] = song
                 self.save_cache_info()
 
             return CacheManager.Result.from_server(
@@ -710,12 +799,12 @@ class CacheManager(metaclass=Singleton):
             return CacheManager.create_future(do_batch_download_songs)
 
         def get_cover_art_filename(
-                self,
-                id: str,
-                before_download: Callable[[], None] = lambda: None,
-                size: Union[str, int] = 200,
-                force: bool = False,
-                allow_download: bool = True,
+            self,
+            id: str,
+            before_download: Callable[[], None] = lambda: None,
+            size: Union[str, int] = 200,
+            force: bool = False,
+            allow_download: bool = True,
         ) -> 'CacheManager.Result[Optional[str]]':
             tag = 'tag_' if self.browse_by_tags else ''
             return self.return_cached_or_download(
@@ -732,13 +821,14 @@ class CacheManager(metaclass=Singleton):
                 before_download: Callable[[], None] = lambda: None,
                 force: bool = False,
         ) -> 'CacheManager.Result[Child]':
-            if self.cache['song_details'].get(song_id) and not force:
+            cache_name = self.id3ify('song_details')
+            if self.cache[cache_name].get(song_id) and not force:
                 return CacheManager.Result.from_data(
-                    self.cache['song_details'][song_id])
+                    self.cache[cache_name][song_id])
 
             def after_download(song_details):
                 with self.cache_lock:
-                    self.cache['song_details'][song_id] = song_details
+                    self.cache[cache_name][song_id] = song_details
                 self.save_cache_info()
 
             return CacheManager.Result.from_server(
@@ -751,10 +841,10 @@ class CacheManager(metaclass=Singleton):
             return CacheManager.create_future(self.server.get_play_queue)
 
         def save_play_queue(
-                self,
-                play_queue: List[str],
-                current: str,
-                position: float,
+            self,
+            play_queue: List[str],
+            current: str,
+            position: float,
         ):
             CacheManager.create_future(
                 self.server.save_play_queue, play_queue, current, position)
@@ -798,6 +888,72 @@ class CacheManager(metaclass=Singleton):
                 before_download=before_download,
                 after_download=after_download,
             )
+
+        def search(
+            self,
+            query,
+            search_callback: Callable[[SearchResult, bool], None],
+            before_download: Callable[[], None] = lambda: None,
+        ):
+            if query == '':
+                search_callback(SearchResult(''), True)
+                return
+
+            # Server search function
+            def server_search():
+                # TODO: if offline, return immediately
+
+                search_fn = (
+                    self.server.search3
+                    if self.browse_by_tags else self.server.search2)
+                result = search_fn(query)
+                yield ('album', result.album)
+                yield ('artist', result.artist)
+                yield ('song', result.song)
+
+            # Local cache search functions
+            def album_search():
+                yield (
+                    'album',
+                    itertools.chain(
+                        *self.cache[self.id3ify('albums')].values()))
+
+            def artist_search():
+                yield ('artist', self.cache[self.id3ify('artists')])
+
+            def song_search():
+                yield (
+                    'song', self.cache[self.id3ify('song_details')].values())
+
+            def playlist_search():
+                yield ('playlist', self.cache['playlists'])
+
+            before_download()
+
+            def do_search():
+                search_future_fns = [
+                    server_search,
+                    album_search,
+                    artist_search,
+                    song_search,
+                    playlist_search,
+                ]
+                search_result = SearchResult(query)
+
+                for i, f in enumerate(as_completed(map(
+                        CacheManager.create_future, search_future_fns))):
+                    try:
+                        for member, result in f.result():
+                            search_result.add_results(member, result)
+                    except Exception as e:
+                        print(e)
+
+                    search_callback(
+                        search_result,
+                        i == len(search_future_fns) - 1,
+                    )
+
+            CacheManager.create_future(do_search)
 
         def get_cached_status(self, song: Child) -> SongCacheStatus:
             cache_path = self.calculate_abs_path(song.path)